tower place holder

+++ app/src/main/java/com/uet/towerdefense/util/Point.java
@@ -1,30 +1,30 @@
 package com.uet.towerdefense.util;
 
 public class Point {
-  private int x, y;
+  private double x, y;
 
   private Point() {
 
   }
 
-  public Point(int x, int y) {
+  public Point(double x, double y) {
     this.x = x;
     this.y = y;
   }
 
-  public int getX() {
+  public double getX() {
     return x;
   }
 
-  public void setX(int x) {
+  public void setX(double x) {
     this.x = x;
   }
 
-  public int getY() {
+  public double getY() {
     return y;
   }
 
-  public void setY(int y) {
+  public void setY(double y) {
     this.y = y;
   }
 }
